增加表格的数据容错提醒
