+++ src/App.vue
@@ -3,58 +3,58 @@
 <template>
     <div class="demo">
         <h2>基础使用,渲染表头和表格内容</h2>
-        <simple-table :data="demo1.data" :columns="demo1.columns"> </simple-table>
+        <lxy-table :data="demo1.data" :columns="demo1.columns"> </lxy-table>
     </div>
     <div class="demo">
         <h2>表格配合开启分页使用（分页事件触发app.vue中的loadData函数）</h2>
-        <simple-table
+        <lxy-table
             v-model:currentPage="demo2.currentPage"
             :data="demo2.data"
             :columns="demo2.columns"
             :pages="demo2.pages"
             @change-page="loadData"
         >
-        </simple-table>
+        </lxy-table>
     </div>
     <div class="demo">
         <h2>表格开启操作列，操作列自定义插槽</h2>
-        <simple-table :data="demo3.data" :open-option="demo3.openOption" :columns="demo3.columns">
+        <lxy-table :data="demo3.data" :open-option="demo3.openOption" :columns="demo3.columns">
             <template v-slot:header-options="scope">
                 <div>自定义操作列表头{{ scope }}</div>
             </template>
             <template v-slot:options="scope">
                 <div>自定义操作列内容{{ scope }}</div>
             </template>
-        </simple-table>
+        </lxy-table>
     </div>
     <div class="demo">
         <h2>自定义渲染表头和表格内容</h2>
-        <simple-table :data="demo3.data" :columns="demo3.columns">
+        <lxy-table :data="demo3.data" :columns="demo3.columns">
             <template v-slot:header-name="scope">
                 <div>自定义name列表头{{ scope }}</div>
             </template>
             <template v-slot:name="scope">
                 <div>自定义name列内容{{ scope }}</div>
             </template>
-        </simple-table>
+        </lxy-table>
     </div>
     <div class="demo">
         <h2>按需开启表格列排序，使用默认排序</h2>
-        <simple-table :data="demo4.data" :columns="demo4.columns"> </simple-table>
+        <lxy-table :data="demo4.data" :columns="demo4.columns"> </lxy-table>
     </div>
     <div class="demo">
         <h2>按需开启表格列排序，使用用户自定义排序(调用app.vue中的onSort自定义排序函数)</h2>
-        <simple-table :data="demo4.data" :on-sort="onSort" :columns="demo4.columns"> </simple-table>
+        <lxy-table :data="demo5.data" :on-sort="onSort" :columns="demo4.columns"> </lxy-table>
     </div>
   <div class="demo">
     <h2>data为空时候</h2>
-    <simple-table  :on-sort="onSort" :columns="demo4.columns"> </simple-table>
+    <lxy-table :on-sort="onSort" :columns="demo4.columns"> </lxy-table>
   </div>
 </template>
 
 <script lang="ts" setup>
 import {reactive} from 'vue'
-import SimpleTable from './components/SimpleTable.vue'
+import lxyTable from './components/lxy_table.vue'
 
 const data = [
     {id: 1, name: 'Liz Lemon', age: 36, is_manager: true, start_date: '02-28-1999', date: new Date('1998-05-16')},
@@ -110,18 +110,29 @@ const demo4 = reactive({
     ],
     data,
 })
+const demo5 = reactive({
+  columns: [
+    {value: 'name', label: 'Name', sort: true},
+    {value: 'age', label: 'Age', sort: true},
+    {value: 'is_manager', label: 'Manager', sort: true},
+    {value: 'start_date', label: 'Start Date', sort: true},
+    {value: 'date', label: '时间', sort: true},
+  ],
+  data,
+})
 
 /**
  * 分页触发事件
  */
 function loadData(page: number) {
     console.log('跳转至', page)
+    demo2.data = demo1.data.slice(Math.random() * 10)
 }
 
 function onSort(data: [], options: any) {
     console.log('传入要排序的list', data)
     console.log('用户点了那一列，正序还是倒序信息', options)
-    return data
+    return data.reverse()
 }
 </script>
