+++ test/SimpleTable.test.ts
@@ -1,6 +1,6 @@
 import { shallowMount } from "@vue/test-utils";
 import { expect, test } from "vitest";
-import SimpleTable from "../src/components/SimpleTable.vue";
+import SimpleTable from "../src/components/lxy_table.vue";
 
 test("mount component", () => {
   const wrapper = shallowMount(SimpleTable, {
